Добавил процессор, сделал квадратку работающей, удалил ненужные файлы, поменял принтф в стеке

+++ Kvadratka/Kvadratka.h
@@ -4,5 +4,5 @@ int aCondition(double a);
 int bCondition(double a, double b, double c, double *x1, double *x2);
 int bcCondition(double b, double c, double *x1, double *x2);
 int cCondition(double a, double b, double c, double *x1, double *x2);
-int Discriminant(double a, double b, double c);
+double Discriminant(double a, double b, double c);
 int kvUr(double a, double b, double c, double *x1, double *x2);
